Rewrite logic for save and return to clinic
Previously, we were handling saves and return to clinic based on
button presses of "save" and "return to clinic", as well as in
state machine of Android (e.g., onPause). However, this made some
things a little inflexible. While the above are still used, the
code is now based on publish and subscribe model where a fragment
registers a save and checkout listener and the main activity will
call these listeners when time to save, checkout, or manage a state
change like pause.

Also bumped the version number.